+++ astropy/table/tests/test_pickle.py
@@ -4,7 +4,9 @@ import numpy as np
 import pytest
 
 from ...table import Table, Column, MaskedColumn, QTable
-from ...units import Quantity
+from ...units import Quantity, deg
+from ...time import Time
+from ...coordinates import SkyCoord
 
 
 def test_pickle_column(protocol):
@@ -48,18 +50,25 @@ def test_pickle_qtable(protocol):
     a = Column(data=[1, 2], name='a', format='%05d', description='col a', unit='cm', meta={'a': 1})
     b = Column(data=[3.0, 4.0], name='b', format='%05d', description='col b', unit='cm',
                meta={'b': 1})
-    t = QTable([a, b], meta={'a': 1})
-    c = Quantity([1, 2], unit='m')
-    t['c'] = c
+    t = QTable([a, b], meta={'a': 1, 'b':Quantity(10,unit='s')})
+    t['c'] = Quantity([1, 2], unit='m')
+    t['d'] = Time(['2001-01-02T12:34:56', '2001-02-03T00:01:02'])
+    t['e'] = SkyCoord([125.0,180.0]*deg, [-45.0,36.5]*deg)
     ts = pickle.dumps(t)
     tp = pickle.loads(ts)
 
     assert np.all(tp['a'] == t['a'])
     assert np.all(tp['b'] == t['b'])
+    # test mixin columns
     assert np.all(tp['c'] == t['c'])
-    assert tp['a'].attrs_equal(t['a'])
-    assert tp['b'].attrs_equal(t['b'])
+    assert np.all(tp['d'] == t['d'])
+    assert np.all(tp['e'].ra == t['e'].ra)
+    assert np.all(tp['e'].dec == t['e'].dec)
+    assert type(tp['c']) == type(t['c'])
+    assert type(tp['d']) == type(t['d'])
+    assert type(tp['e']) == type(t['e'])
     assert tp.meta == t.meta
+    assert type(tp) == type(t)
 
 def test_pickle_masked_table(protocol):
     a = Column(data=[1, 2], name='a', format='%05d', description='col a', unit='cm', meta={'a': 1})